+++ CHANGELOG.md
@@ -4,12 +4,13 @@ X.Y.Z Release notes
 * None
 
 ### Enhancements
+* Added methods `Realm.beginTransaction()`, `Realm.commitTransaction()`, `Realm.cancelTransaction()` to manually control write transactions.
+* Added property `Realm.isInTransaction` which indicates if write transaction is in progress.
 * Added `shouldCompactOnLaunch` to configuration (#507).
 * Added `Realm.compact()` for manually compacting Realm files.
 
 ### Bug fixes
 * None
-
 
 1.10.3 Release notes (2017-8-16)
 =============================================================
@@ -20,7 +21,7 @@ X.Y.Z Release notes
 * None
 
 ### Bug fixes
-* none
+* None
 
 
 1.10.2 Release notes (2017-8-16)